+++ src/components/organism/post/BoardTypeSelecSection.tsx
