+++ src/components/organism/post/EditorSection.tsx
@@ -1,11 +1,10 @@
 import styled from 'styled-components';
 import dynamic from 'next/dynamic';
 import { BlockNoteEditor } from '@blocknote/core';
+import { usePopup } from '@src/hooks';
 import { ERROR_MESSAGE } from '@src/constants/message';
 import { POPUP_DURATION } from '@src/constants/duration';
-import { PostReqData } from '@src/types/PostTypes';
 import { useFormContext } from 'react-hook-form';
-import { usePopup } from '@src/hooks';
 
 const BlockNote = dynamic(
   () => import('@src/components/molecules/editor/BlockNote'),
@@ -15,7 +14,7 @@ const BlockNote = dynamic(
 );
 
 function EditorSection() {
-  const { setValue } = useFormContext<PostReqData>();
+  const { setValue } = useFormContext();
   const { openAndDeletePopup } = usePopup();
 
   const onEditorContentChange = async (editor: BlockNoteEditor) => {
